requirements.txt
django==4.2.6
opa-python-client==1.3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
gunicorn==21.2.0


